# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Storecipe public API contract
  version: 0.3.0
  description: >-
    Public contract for recipes, ratings, deterministic recipe queries, and imports. The internal
    `POST /internal/recipes/source-lookup` endpoint is M2M-only and is excluded
    from the public proxy.
servers:
  - url: https://api.storecipe.example
security:
  - bearerAuth: []
paths:
...
  /v1/recipes:
    get:
      operationId: queryRecipes
      description: Requires the `recipes:read` scope.
      parameters:
        - {name: text, in: query, schema: {type: [string, "null"], maxLength: 200}}
        - name: ingredient
          in: query
          style: form
          explode: true
          description: Every listed value is required (AND).
          schema:
            type: array
            maxItems: 32
            items: {type: string, minLength: 1, maxLength: 200}
        - name: tag
          in: query
          style: form
          explode: true
          description: Every listed value is required (AND).
          schema:
            type: array
            maxItems: 16
            items: {type: string, minLength: 1, maxLength: 64}
        - {name: maxTotalMinutes, in: query, schema: {type: [integer, "null"], minimum: 0}}
        - {name: minRating, in: query, schema: {type: [integer, "null"], minimum: 1, maximum: 5}}
        - name: ratingState
          in: query
          schema: {type: string, enum: [any, rated, unrated], default: any}
        - name: favorite
          in: query
          description: Favorites-only filter. Omit for all recipes; `false` is invalid.
          schema: {type: boolean, const: true}
        - name: sort
          in: query
          style: form
          explode: true
          description: Ordered sort precedence. The service appends its private recipeId tie-breaker.
          examples:
            ratingFirst:
              value: [rating:desc, totalMinutes:asc]
          schema:
            type: array
            maxItems: 6
            items: {$ref: "#/components/schemas/RecipeSort"}
        - {name: cursor, in: query, schema: {type: [string, "null"], maxLength: 1024}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        "200":
          description: User-owned recipes
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RecipeQueryPage"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/ValidationError"}
        "409": {$ref: "#/components/responses/StaleRecipeQueryCursor"}
        "410": {$ref: "#/components/responses/AccountDeleted"}
        "414": {$ref: "#/components/responses/UriTooLong"}
    post:
      operationId: createRecipe
      description: Requires the `recipes:write` scope.
      parameters:
        - {$ref: "#/components/parameters/RecipeCreateIdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RecipeCreate"}
      responses:
        "201":
          description: Created recipe
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Recipe"}
        "200":
          description: Existing idempotent recipe replay
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Recipe"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/IdempotencyConflict"}
        "410": {$ref: "#/components/responses/AccountDeleted"}
        "413": {$ref: "#/components/responses/RequestTooLarge"}
        "422": {$ref: "#/components/responses/ValidationError"}
        "429": {$ref: "#/components/responses/CatalogRateLimited"}
        "503": {$ref: "#/components/responses/RateLimitUnavailable"}
  /v1/recipes/{recipeId}:
    parameters:
      - {$ref: "#/components/parameters/RecipeId"}
    get:
      operationId: getRecipe
      description: Requires `recipes:read`; another user's recipe is reported as not found.
      responses:
        "200": {description: Recipe, content: {application/json: {schema: {$ref: "#/components/schemas/Recipe"}}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "410": {$ref: "#/components/responses/AccountDeleted"}
    patch:
      operationId: updateRecipe
      description: Requires the `recipes:write` scope.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RecipePatch"}
      responses:
        "200": {description: Updated recipe, content: {application/json: {schema: {$ref: "#/components/schemas/Recipe"}}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "410": {$ref: "#/components/responses/AccountDeleted"}
        "413": {$ref: "#/components/responses/RequestTooLarge"}
        "422": {$ref: "#/components/responses/ValidationError"}
        "429": {$ref: "#/components/responses/CatalogRateLimited"}
        "503": {$ref: "#/components/responses/RateLimitUnavailable"}
    delete:
      operationId: deleteRecipe
      description: Requires the `recipes:write` scope.
      responses:
        "204": {description: Deleted}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "410": {$ref: "#/components/responses/AccountDeleted"}
        "429": {$ref: "#/components/responses/CatalogRateLimited"}
        "503": {$ref: "#/components/responses/RateLimitUnavailable"}
  /v1/recipes/{recipeId}/cooked:
    parameters:
      - {$ref: "#/components/parameters/RecipeId"}
    post:
      operationId: markRecipeCooked
      description: Requires `recipes:write`; records the server time as the last cooked time.
      responses:
        "200": {description: Updated recipe, content: {application/json: {schema: {$ref: "#/components/schemas/Recipe"}}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "410": {$ref: "#/components/responses/AccountDeleted"}
        "429": {$ref: "#/components/responses/CatalogRateLimited"}
        "503": {$ref: "#/components/responses/RateLimitUnavailable"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RecipeId:
      name: recipeId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    RecipeCreateIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$'}
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      headers:
        WWW-Authenticate:
          schema: {type: string}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Forbidden:
      description: Access token lacks a required scope
      headers:
        WWW-Authenticate:
          schema: {type: string}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    AccountDeleted:
      description: The authenticated subject is protected by a deletion tombstone.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
          example:
            type: https://docs.storecipe.example/problems/account_deleted
            title: Gone
            status: 410
            detail: This account has been deleted.
            request_id: 01JEXAMPLEACCOUNTDELETED
            errorCategory: account_deleted
    NotFound:
      description: Resource not found or not owned by the authenticated user
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    UriTooLong:
      description: Raw query string exceeds 6144 bytes
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    IdempotencyConflict:
      description: Idempotency key was already used for different recipe content
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
          example:
            type: https://docs.storecipe.example/problems/idempotency_conflict
            title: Conflict
            status: 409
            detail: Idempotency key was already used for different recipe content.
            request_id: 01JEXAMPLEIDEMPOTENCY
            errorCategory: idempotency_conflict
    StaleRecipeQueryCursor:
      description: The catalog changed after the referenced recipe-query page was read.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
          example:
            type: https://docs.storecipe.example/problems/stale_recipe_query_cursor
            title: Conflict
            status: 409
            detail: Recipe query cursor is stale.
            request_id: 01JEXAMPLESTALECURSOR
            errorCategory: stale_recipe_query_cursor
...
    CatalogRateLimited:
      description: Authenticated Catalog mutation burst limit exceeded; retry after the indicated interval.
      headers:
        Retry-After:
          schema: {type: integer, minimum: 1}
        RateLimit-Limit:
          schema: {type: integer, minimum: 1}
        RateLimit-Remaining:
          schema: {type: integer, minimum: 0}
        RateLimit-Reset:
          schema: {type: integer, minimum: 0}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    RateLimitUnavailable:
      description: Redis could not make a rate-limit admission decision; retry after the indicated interval.
      headers:
        Retry-After:
          schema: {type: integer, minimum: 1}
        RateLimit-Limit:
          schema: {type: integer, minimum: 1}
        RateLimit-Remaining:
          schema: {type: integer, minimum: 0}
        RateLimit-Reset:
          schema: {type: integer, minimum: 0}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    RequestTooLarge:
      description: Request body exceeds the service size limit.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
  schemas:
...
    RecipeCreate:
      type: object
      required: [title, ingredients, instructions]
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        sourceUrl: {type: [string, "null"], format: uri, minLength: 1, maxLength: 2083, pattern: '^https?://[^/?#]+(?:[/?#].*)?$', description: 'HTTP(S) URL with a host; Catalog normalizes a host-only URL with a trailing slash.'}
        servings: {type: [integer, "null"], minimum: 1, maximum: 2147483647}
        prepMinutes: {type: [integer, "null"], minimum: 0, maximum: 2147483647}
        cookMinutes: {type: [integer, "null"], minimum: 0, maximum: 2147483647}
        totalMinutes: {type: [integer, "null"], minimum: 0, maximum: 2147483647}
        ingredients:
          type: array
          minItems: 1
          maxItems: 256
          items: {$ref: "#/components/schemas/Ingredient"}
        instructions:
          type: array
          minItems: 1
          maxItems: 256
          items: {type: string, minLength: 1, maxLength: 4096}
        tags:
          type: array
          maxItems: 64
          items: {type: string, minLength: 1, maxLength: 64}
          default: []
    RecipePatch:
      type: object
      minProperties: 1
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        sourceUrl: {type: [string, "null"], format: uri, minLength: 1, maxLength: 2083, pattern: '^https?://[^/?#]+(?:[/?#].*)?$', description: 'HTTP(S) URL with a host; Catalog normalizes a host-only URL with a trailing slash.'}
        servings: {type: [integer, "null"], minimum: 1, maximum: 2147483647}
        prepMinutes: {type: [integer, "null"], minimum: 0, maximum: 2147483647}
        cookMinutes: {type: [integer, "null"], minimum: 0, maximum: 2147483647}
        totalMinutes: {type: [integer, "null"], minimum: 0, maximum: 2147483647}
        ingredients:
          type: array
          minItems: 1
          maxItems: 256
          items: {$ref: "#/components/schemas/Ingredient"}
        instructions:
          type: array
          minItems: 1
          maxItems: 256
          items: {type: string, minLength: 1, maxLength: 4096}
        tags:
          type: array
          maxItems: 64
          items: {type: string, minLength: 1, maxLength: 64}
        favorite: {type: boolean}
        personalNotes: {type: [string, "null"], maxLength: 5000}
    CoverImage:
      type: object
      required: [url, etag, byteSize, contentType]
      properties:
        url: {type: string, minLength: 1}
        etag: {type: string, pattern: '^[0-9a-f]{64}$'}
        byteSize: {type: integer, exclusiveMinimum: 0, maximum: 1572864}
        contentType: {type: string, enum: [image/webp]}
    Recipe:
      type: object
      required: [id, title, sourceUrl, servings, prepMinutes, cookMinutes, totalMinutes, ingredients, instructions, tags, rating, coverImage, favorite, personalNotes, lastCookedAt]
      properties:
        id: {type: string, format: uuid}
        title: {type: string, minLength: 1, maxLength: 200}
        sourceUrl: {type: [string, "null"], format: uri, minLength: 1, maxLength: 2083, pattern: '^https?://[^/?#]+(?:[/?#].*)?$', description: 'HTTP(S) URL with a host; Catalog normalizes a host-only URL with a trailing slash.'}
        servings: {type: [integer, "null"], minimum: 1}
        prepMinutes: {type: [integer, "null"], minimum: 0}
        cookMinutes: {type: [integer, "null"], minimum: 0}
        totalMinutes: {type: [integer, "null"], minimum: 0}
        ingredients: {type: array, items: {$ref: "#/components/schemas/Ingredient"}}
        instructions: {type: array, items: {type: string, minLength: 1}}
        tags: {type: array, items: {type: string}}
        rating: {type: [integer, "null"], minimum: 1, maximum: 5}
        coverImage:
          anyOf:
            - {$ref: "#/components/schemas/CoverImage"}
            - {type: "null"}
        favorite: {type: boolean}
        personalNotes: {type: [string, "null"], maxLength: 5000}
        lastCookedAt: {type: [string, "null"], format: date-time}
    Ingredient:
      type: object
      required: [rawText, name, canonicalName]
      properties:
        rawText: {type: string, minLength: 1, maxLength: 4096}
        name: {type: string, minLength: 1, maxLength: 200}
        canonicalName: {type: string, minLength: 1, maxLength: 200}
        quantity: {type: [number, "null"], minimum: 0}
        unit: {type: [string, "null"], minLength: 1, maxLength: 64}
    RecipeQueryPage:
      type: object
      required: [items, nextCursor]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/Recipe"}}
        nextCursor: {type: [string, "null"]}
    RecipeSort:
      type: string
      enum:
        - rating:asc
        - rating:desc
        - totalMinutes:asc
        - totalMinutes:desc
        - createdAt:asc
        - createdAt:desc
        - updatedAt:asc
        - updatedAt:desc
        - title:asc
        - title:desc
...
    Problem:
      type: object
      required: [type, title, status, request_id]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        request_id: {type: string}
        errorCategory: {type: string}
        existingJobId: {type: string, format: uuid}
        existingRecipeId: {type: string, format: uuid}